Allow integer orderReference to be auto generated

+++ src/main/java/com/sullivan/ricksbricks/repository/BrickOrdersRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BrickOrdersRepository extends JpaRepository<BrickOrderEntity, Long> {
+public interface BrickOrdersRepository extends JpaRepository<BrickOrderEntity, Integer> {
 }
